the-hirst-painting-project/requirements.txt
colorgram.py==1.2.0
